0.1.0
병합 버튼 추가 (이미지를 하나의 PDF로 병합시켜줌)
초기화 버튼 추가
progress bar 추가
자잘한 예외처리 및 사용성/편의성 올리기 필살기
